Rename plants class file
