+++ plants.rb
